chore: outsource functions to own utils file
